+++ cmake/BLTOptions.cmake
@@ -1,4 +1,4 @@
-# Copyright (c) 2017-2021, Lawrence Livermore National Security, LLC and
+# Copyright (c) 2017-2022, Lawrence Livermore National Security, LLC and
 # other BLT Project Developers. See the top-level LICENSE file for details
 # 
 # SPDX-License-Identifier: (BSD-3-Clause)
@@ -59,13 +59,8 @@ mark_as_advanced(ENABLE_CLANG_CUDA)
 set(BLT_CLANG_CUDA_ARCH "sm_30" CACHE STRING "Compute architecture to use when generating CUDA code with Clang")
 mark_as_advanced(BLT_CLANG_CUDA_ARCH)
 option(ENABLE_HIP         "Enable HIP support" OFF)
-cmake_dependent_option(ENABLE_CLANG_HIP   "Enable Clang's native HIP support" OFF
-                      "ENABLE_HIP" OFF)
-mark_as_advanced(ENABLE_CLANG_HIP)
-set(BLT_CLANG_HIP_ARCH "gfx906" CACHE STRING "AMDGPU ISA target to use when generating HIP code with Clang" )
-mark_as_advanced(BLT_CLANG_HIP_ARCH)
-option(ENABLE_HCC         "Enable HCC support" OFF)
-set(BLT_ROCM_ARCH "gfx900" CACHE STRING "gfx architecture to use when generating ROCm code")
+option(ENABLE_HCC "Enable HCC support" OFF)
+set(CMAKE_HIP_ARCHITECTURES "gfx900" CACHE STRING "gfx architecture to use when generating HIP/ROCm code")
 
 #------------------------------------------------------------------------------
 # Test Options